backend/requirements.txt
fastapi
uvicorn

#for template
jinja2

#for static files
aiofiles

#for database
sqlalchemy
psycopg2

#for loading environment variables
python-dotenv